fix(rope): use fp32 for sin/cos accumulation to handle extreme-long seq (G7.8c)
ROPE fp16 sin/cos computation overflows for extreme sequence lengths (N >=
65504 -> inf, N >= 2049 -> integer rounding), producing NaN outputs that the
correctness probe legitimately flagged as 'mismatch'. PyTorch's reference
implementation uses fp32 internally then casts to the input dtype; we now
mirror that contract across all five rope implementations in the harness.

Five sites patched (all using torch.arange + inv_freq + sin/cos pattern):

  1. benchmarks/bench_l1.py::_torch_reference         (rope golden)
  2. benchmarks/baselines/arke_runner.py::run_with_inputs       (rope path)
  3. benchmarks/baselines/arke_runner.py::_fill_missing_inputs  (rope inputs)
  4. benchmarks/baselines/pytorch_eager.py::get_fn              (closure)
  5. benchmarks/baselines/pytorch_eager.py::run_with_inputs     (independent)

Site #5 was the one missed in the prior pass; it is the path the harness
runner-baseline invokes for rope correctness probes, and was the source of
the lingering shape11 (1,128,64) max_abs_diff=0.148 discrepancy: arke ===
_torch_reference (diff=0.0), but pytorch_eager.run_with_inputs differed by
0.148 because it ran its own fp16 sin/cos.

Computation pattern in all five sites (and matching PyTorch reference):

    inv_freq = 1.0 / (base ** (arange(0, D, 2, dtype=fp32) / D))
    freqs    = einsum('i,j->ij', arange(S, dtype=fp32), inv_freq)
    cos, sin = freqs.cos().to(x.dtype), freqs.sin().to(x.dtype)

Verification (tier-3, all 16 rope shapes):
  - 48 rows total, 0 mismatch
  - extreme-long (32, 65536, 128): arke 254ms vs Liger 628ms (2.5x faster);
    pytorch-eager preflight-skipped, audit-row only (golden_unavailable)
  - qwen25-32k (28, 32768, 128): allclose=true, max_abs=0
  - llama3-8k (16, 8192, 128):   allclose=true, max_abs=0
  - non-align-1/2 (odd D):        all three runners typed-unsupported (correct)
  - 9 mid shapes:                 allclose=true, max_abs=0
  - extreme-batch:                allclose=true, max_abs=0

Unit tests: 53 passed, 1 skipped (test_arke_runner_correctness_probe,
test_gate_g7, test_bench_l1_smoke, test_benchmark_correctness_probe_linea12,
test_ssot_op_registry).

G7.8c correctness failure count: 5 -> 1 (rope no longer in failure list).

Refs: docs/benchmark/golden-kernel-ladder.md, AGENTS.md (fp16 integer
precision boundary at N=2049/65504).

+++ benchmarks/baselines/arke_runner.py
@@ -188,14 +188,21 @@ class ArkeRunner(BaselineRunner):
                 x = filled[0]
                 head_dim = x.shape[-1]
                 seq_len = x.shape[-2]
+                # Numerical-stability fix (2026-05-14, Q5a): compute sin/cos
+                # in fp32 then cast back. fp16 torch.arange(seq) loses integer
+                # precision at seq>=2048 and overflows past fp16_max≈65504, so
+                # at extreme-long (seq=65536) freqs → NaN and the rope output
+                # collapses to NaN. Same fix mirrored in bench_l1._torch_reference
+                # and baselines/pytorch_eager.py. Matches HF Transformers /
+                # Liger-Kernel / FlashInfer rotary-frequency convention.
                 freqs = torch.einsum(
                     "i,j->ij",
-                    torch.arange(seq_len, device=x.device, dtype=x.dtype),
-                    1.0 / (10000 ** (torch.arange(0, head_dim, 2, device=x.device, dtype=x.dtype) / head_dim)),
+                    torch.arange(seq_len, device=x.device, dtype=torch.float32),
+                    1.0 / (10000 ** (torch.arange(0, head_dim, 2, device=x.device, dtype=torch.float32) / head_dim)),
                 )
                 emb = torch.cat([freqs, freqs], dim=-1)
-                cos_emb = torch.cos(emb).unsqueeze(0)
-                sin_emb = torch.sin(emb).unsqueeze(0)
+                cos_emb = torch.cos(emb).unsqueeze(0).to(x.dtype)
+                sin_emb = torch.sin(emb).unsqueeze(0).to(x.dtype)
                 x1 = x[..., : head_dim // 2]
                 x2 = x[..., head_dim // 2 :]
                 rotated = torch.cat([-x2, x1], dim=-1)
@@ -246,12 +253,21 @@ class ArkeRunner(BaselineRunner):
         if op == "rope" and len(inputs) == 1:
             head_dim = x.shape[-1]
             seq_len = x.shape[-2]
+            # Numerical-stability fix (2026-05-14, Q5a): compute the rotary
+            # frequencies in fp32 then cast cos/sin back to the input dtype.
+            # fp16 torch.arange(seq) loses integer precision at seq>=2048 and
+            # overflows past fp16_max≈65504, so for extreme-long (seq=65536)
+            # freqs decays to NaN/inf and downstream SemanticInterpreter
+            # produces a NaN tensor. This path FEEDS THE SEMANTIC INTERPRETER —
+            # the (cos, sin) tensors it synthesizes are what ref_rope multiplies
+            # against the fp16 x. Matches HF Transformers / Liger-Kernel /
+            # FlashInfer rotary-frequency convention (fp32 trig, fp16 hadamard).
             freqs = torch.einsum(
                 "i,j->ij",
-                torch.arange(seq_len, device=device, dtype=dtype),
-                1.0 / (10000 ** (torch.arange(0, head_dim, 2, device=device, dtype=dtype) / head_dim)),
+                torch.arange(seq_len, device=device, dtype=torch.float32),
+                1.0 / (10000 ** (torch.arange(0, head_dim, 2, device=device, dtype=torch.float32) / head_dim)),
             )
-            return [x, torch.cos(freqs), torch.sin(freqs)]
+            return [x, torch.cos(freqs).to(dtype), torch.sin(freqs).to(dtype)]
         return inputs
 
     @staticmethod
